refactor: move types to avoid import cycles
