+++ pkg/xds/envoy/tags/match.go
@@ -10,7 +10,6 @@ import (
 
 	common_api "github.com/kumahq/kuma/api/common/v1alpha1"
 	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
-	core_policy "github.com/kumahq/kuma/pkg/core/policy"
 	"github.com/kumahq/kuma/pkg/util/maps"
 )
 
@@ -255,41 +254,4 @@ func TagKeySlice(tags []Tags) TagKeysSlice {
 		r[i] = tags[i].Keys()
 	}
 	return r
-}
-
-func MatchingRegex(tags mesh_proto.SingleValueTagSet) string {
-	var re string
-	for _, key := range tags.Keys() {
-		keyIsEqual := fmt.Sprintf(`&%s=`, key)
-		var value string
-		switch tags[key] {
-		case "*":
-			value = ``
-		default:
-			value = fmt.Sprintf(`[^&]*%s[,&]`, tags[key])
-		}
-		value = strings.ReplaceAll(value, ".", `\.`)
-		expr := keyIsEqual + value + `.*`
-		re += expr
-	}
-	re = `.*` + re
-	return re
-}
-
-func RegexOR(r ...string) string {
-	if len(r) == 0 {
-		return ""
-	}
-	if len(r) == 1 {
-		return r[0]
-	}
-	return fmt.Sprintf("(%s)", strings.Join(r, "|"))
-}
-
-func MatchSourceRegex(policy core_policy.ConnectionPolicy) string {
-	var selectorRegexs []string
-	for _, selector := range policy.Sources() {
-		selectorRegexs = append(selectorRegexs, MatchingRegex(selector.Match))
-	}
-	return RegexOR(selectorRegexs...)
 }
